[choose][m]: refactor component to work with import- #64
