+++ package.json
@@ -17,12 +17,14 @@
     "ckanClient": "git+https://github.com/datopian/ckan-client-js.git",
     "data.js": "git+https://github.com/datopian/data.js.git",
     "frictionless-ckan-mapper-js": "^1.0.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-scripts": "^3.4.3",
-    "react-table": "^7.5.0",
     "stream-to-array": "^2.3.0",
-    "uuidv4": "^6.2.3"
+    "uuidv4": "^6.2.3",
+    "react-table": "^7.5.0",
+    "react-scripts": "^3.4.3"
+  },
+  "peerDependencies": {
+    "react": "^16.13.0",
+    "react-dom": "^16.13.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -69,6 +71,7 @@
     "babel-loader": "^8.1.0",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.3",
+    "eslint-plugin-react-hooks": "^4.1.2",
     "react-test-renderer": "^16.13.1"
   },
   "babel": {
