+++ src/lib/App.js
@@ -4,11 +4,11 @@ import PropTypes from "prop-types";
 import frictionlessCkanMapper from "frictionless-ckan-mapper-js";
 import { v4 as uuidv4 } from "uuid";
 
-import Metadata from "./lib/components/Metadata";
-import TableSchema from "./lib/components/TableSchema";
-import Upload from "./lib/components/Upload";
+import Metadata from "./components/Metadata";
+import TableSchema from "./components/TableSchema";
+import Upload from "./components/Upload";
 import "./App.css";
-import { removeHyphen } from "./lib/utils";
+import { removeHyphen } from "./utils";
 
 export class ResourceEditor extends React.Component {
   constructor(props) {
